Fix warning for PHP 8.1 where float converts to int

+++ src/Component/UlidGenerator.php
@@ -71,7 +71,7 @@ class UlidGenerator implements IdGeneratorInterface
      */
     private function getCurrentTime(): int
     {
-        return microtime(true) * 1000;
+        return (int) (microtime(true) * 1000);
     }
 
     /**
